+++ main.go
@@ -21,7 +21,7 @@ import (
 )
 
 //	@title			Device Management Service
-//	@version		0.4.110
+//	@version		0.4.111
 //	@description	A dashboard application for computing providers.
 //	@termsOfService	https://nunet.io/tos